+++ indico/modules/rb_new/controllers/management/events.py
